pyproject.toml
[tool.poetry]
name = "eora"
version = "0.1.0"
description = ""
authors = ["rudich <rudi4.clarinet@mail.ru>"]
readme = "README.md"
packages = [{include = "eora"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.43"
redis = "^6.4.0"
httpx = "^0.28.1"
beautifulsoup4 = "^4.13.4"
openai = "^1.99.9"
python-dotenv = "^1.1.1"
orjson = "^3.11.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
